feat: use multiple editors in one browser tab
