+++ apps/playground/src/App.tsx
@@ -1,21 +1,14 @@
 import '@blocksuite/editor/themes/affine.css';
 import './App.css';
 
-import {Editor} from "./components/Editor.tsx";
-import {Iframe} from "./components/Iframe.tsx";
+// @ts-ignore
+import {WebsocketProvider} from "y-websocket";
+import Editor from "./components/Editor.tsx";
 
 function App() {
     return (
-        <div>
-            {[...Array(2)].map((_, i) => i).map(id => {
-                return <div key={id}>
-                    <Iframe>
-                        <Editor/>
-                    </Iframe>
-                </div>
-            })}
-        </div>
-    );
+        <Editor editorNum={2} />
+    )
 }
 
 export default App
